backend/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.3
distlib==0.3.4
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.78.0
feedparser==6.0.10
filelock==3.7.1
frozenlist==1.3.0
h11==0.13.0
httptools==0.4.0
identify==2.5.1
idna==3.3
motor==3.0.0
multidict==6.0.2
nodeenv==1.6.0
platformdirs==2.5.2
pre-commit==2.19.0
pydantic==1.9.1
pymongo==4.1.1
python-decouple==3.6
python-dotenv==0.20.0
PyYAML==6.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
typing_extensions==4.2.0
uvicorn==0.17.6
uvloop==0.16.0
virtualenv==20.14.1
watchgod==0.8.2
websockets==10.3
yarl==1.7.2
